apptest: Rewrite legacy backup/restore test so that it is supported by cluster
Note that the test fails for cluster when switching from pt index to legacy back
again. When switching like so, the queries should not return data inserted in pt
index. However it is returned for query_range queries in vmsingle and the test
capture this. When the same test is run for cluster, the pt index data is not
returned, which is expected anyway, but this means different behavior in single
node and cluster. Will debug this difference later and fix in follow-up commits.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>